Checkpoint change (part 1): move SyncThread out of Bookie
this change is major moving the code out of Bookie for porting the checkpoint changes from OSS

RB_ID=731972
